+++ src/main/webapp/WEB-INF/configs/spring-config-hibernate.xml
@@ -20,7 +20,7 @@
 			p:driverClassName="com.microsoft.sqlserver.jdbc.SQLServerDriver"
 			p:url="jdbc:sqlserver://localhost:1433;databaseName = AutoPartsDB;encrypt=true;trustServerCertificate=true;"
 			p:username="sa"
-			p:password="minhduc63">
+			p:password="1111">
 	</bean>
 	
 	<!-- Cấu hình Hibernate -->
